dev: move k8s primitives to model
