+++ pkg/controller/chi/poller.go
@@ -26,7 +26,7 @@ import (
 	api "github.com/altinity/clickhouse-operator/pkg/apis/clickhouse.altinity.com/v1"
 	"github.com/altinity/clickhouse-operator/pkg/chop"
 	"github.com/altinity/clickhouse-operator/pkg/controller"
-	"github.com/altinity/clickhouse-operator/pkg/model/chi/creator/primitives"
+	"github.com/altinity/clickhouse-operator/pkg/model/primitives"
 	"github.com/altinity/clickhouse-operator/pkg/util"
 )
