+++ pkg/model/chi/creator/service.go
@@ -23,7 +23,7 @@ import (
 
 	api "github.com/altinity/clickhouse-operator/pkg/apis/clickhouse.altinity.com/v1"
 	model "github.com/altinity/clickhouse-operator/pkg/model/chi"
-	"github.com/altinity/clickhouse-operator/pkg/model/chi/creator/primitives"
+	"github.com/altinity/clickhouse-operator/pkg/model/primitives"
 	"github.com/altinity/clickhouse-operator/pkg/util"
 )
